backend/requirements.txt
flask
flask-cors
beautifulsoup4
requests
gunicorn
psycopg2-binary
